chore(lighthouse): move server to tcp_server
