+++ lighthouse/tcp_server/config.go
@@ -1,4 +1,4 @@
-package server
+package tcp_server
 
 import "crypto/ed25519"
