+++ lighthouse/tcp_server/server.go
@@ -1,4 +1,4 @@
-package server
+package tcp_server
 
 // Server ...
 type Server struct {
