+++ lighthouse/tcp_server/store.go
@@ -1,4 +1,4 @@
-package server
+package tcp_server
 
 import (
 	"errors"
