+++ lighthouse/tcp_server/tcp.go
@@ -1,4 +1,4 @@
-package server
+package tcp_server
 
 import (
 	"fmt"
